Fix option checkout for bit positions over 31

+++ src/client/srcml_options.cpp
@@ -9,4 +9,4 @@
 
 #include <srcml_options.hpp>
 
-int SRCMLOptions::opt;
+uint_fast64_t SRCMLOptions::opt;
