+++ src/client/srcml_options.hpp
@@ -10,30 +10,32 @@
 #ifndef SRCML_OPTIONS_HPP
 #define SRCML_OPTIONS_HPP
 
+#include <cstdint>
+
 class SRCMLOptions {
 public:
-    friend void enable(int option);
+    friend void enable(uint_fast64_t option);
 
-    static void set(int options) {
+    static void set(uint_fast64_t options) {
 
         opt = options;
     }
 
-    static int get()  {
+    static uint_fast64_t get()  {
 
         return opt;
     }
 
  private:
-    static int opt;
+    static uint_fast64_t opt;
 };
 
-inline bool option(int option) {
+inline bool option(uint_fast64_t option) {
 
     return SRCMLOptions::get() & option;
 }
 
-inline void enable(int option) {
+inline void enable(uint_fast64_t option) {
 
     SRCMLOptions::opt |= option;
 }
